requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2024.2
Django==5.0.2
django-bootstrap5==23.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.11
django-crispy-forms==2.1
django-js-asset==2.2.0
django-multiupload==0.6.1
django-tinymce==3.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
